Move get_first_page_of_pdf_as_image fn to base/utils
